Major restructure probably forgot a lot of stuff I will fix later -Divided into 3 major functionality files -Removed obsolete or questionable code -Started refactoring for multi threading
